Be fix 피드 작성자 닉네임, 프로필이미지url 추가 및 이미지태그 좌표값 변경

+++ server/src/main/java/com/mainproject/server/feed/dto/FeedResponseDto.java
@@ -37,8 +37,8 @@ public class FeedResponseDto {
             private String productName;
             private String productPrice;
             private String productInfo;
-            private Long positionX;
-            private Long positionY;
+            private Double positionX;
+            private Double positionY;
         }
     }
 }
